feat: Updates project according to QA task list
