+++ app/data/global-data.ts
@@ -14,7 +14,7 @@ export const accountData = {
     text: 'Salvar alterações',
     variant: 'orange',
   },
-};
+}
 
 // HEADER
 
@@ -78,7 +78,7 @@ export const headerData = {
       blank: false,
     },
   ],
-};
+}
 
 // FOOTER
 
@@ -137,7 +137,7 @@ export const footerData = {
       blank: true,
     },
   ],
-};
+}
 
 // BANK STATEMENT
 
@@ -145,27 +145,27 @@ export const bankStatementData = {
   title: 'Extrato',
   transactions: [
     {
-      date: '2025-06-18',
+      date: '18/06/2025',
       amount: 150,
       type: 'deposit',
     },
     {
-      date: '2025-06-21',
+      date: '21/06/2025',
       amount: 100,
       type: 'deposit',
     },
     {
-      date: '2025-06-21',
+      date: '21/06/2025',
       amount: 50,
       type: 'deposit',
     },
     {
-      date: '2025-06-21',
+      date: '21/06/2025',
       amount: 500,
       type: 'transfer',
     },
   ],
-};
+}
 
 // 404
 
@@ -178,7 +178,7 @@ export const notFoundData = {
     url: '/',
     blank: false,
   },
-};
+}
 
 // HOME LOGGED-OUT
 
@@ -212,7 +212,14 @@ export const loggedOutData = {
         'Seus dispositivos móveis (computador e laptop) protegidos por uma mensalidade simbólica.',
     },
   ],
-};
+}
+
+// DAHBOARD HERO
+
+export const dashboardHeroData = {
+  amountLabel: 'Saldo',
+  accountLabel: 'Conta corrente',
+}
 
 // TRANSACTIONS
 
@@ -232,7 +239,7 @@ export const transactionsData = {
   ],
   imageAlt:
     'Imagem de mulher com cartão de crédito. A mulher está segurando o cartão com a mão esquerda e está olhando para o lado direito. A imagem está em um fundo cinza.',
-};
+}
 
 // INVESTMENTS
 
@@ -250,7 +257,7 @@ export const investmentsData = {
       price: 14000,
     },
   ],
-};
+}
 
 // SERVICES
 
@@ -282,4 +289,4 @@ export const servicesData = {
       type: 'cellphone',
     },
   ],
-};
+}
