+++ app/layouts/blocks/account/account-form.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { IAccount } from './account'
-import Cta from '../cta'
+import Cta from '../../../components/cta'
 import { useRouter } from 'next/router'
-import Input from '../input'
+import Input from '../../../components/input'
 
 const AccountForm: React.FC<IAccount> = ({
   cta,
