+++ app/layouts/blocks/investments/investment-chart.tsx
