+++ app/layouts/blocks/investments/investment-rate-card.tsx
