+++ app/layouts/blocks/transactions/transaction-card-layout.tsx
@@ -7,11 +7,7 @@ const TransactionsCardLayout = ({
 }) => {
   return (
     <section
-      className="flex flex-col relative
-        pb-[141px]
-        rounded-lg overflow-hidden
-        min-h-[633px] md:min-h-[629px] lg:min-h-100
-        p-4 md:p-8 lg:p-8"
+      className="flex flex-col relative pb-[8.8125rem] rounded-lg overflow-hidden min-h-[39.5625rem] md:min-h-[39.3125rem] lg:min-h-100 p-4 md:p-8 lg:p-8 bg-gray-medium"
     >
       {children}
     </section>
