+++ app/package.json
@@ -11,6 +11,7 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "date-fns": "^2.30.0",
     "next": "15.3.3",
     "next-plugin-svgr": "^1.1.10",
     "next-seo": "^5.5.0",
