+++ app/pages/investments.tsx
@@ -1,9 +1,9 @@
-import AsideNav from "@/components/aside-nav"
-import DashboardHero from "@/components/dashboard-hero"
-import BankStatement from "@/components/bank-statement"
+import AsideNav from "@/layouts/blocks/aside-nav"
+import DashboardHero from "@/layouts/blocks/dashboard-hero"
+import BankStatement from "@/layouts/blocks/bank-statement"
 import { NextSeo } from "next-seo"
 import useStateController from "@/hooks/use-state-controller"
-import InvestimentCard from "@/components/investments/investment-card"
+import InvestimentCard from "@/layouts/blocks/investments/investment-card"
 
 export default function InvestmentsPage() {
   const {authStatus} = useStateController()
@@ -16,9 +16,9 @@ export default function InvestmentsPage() {
           <AsideNav />
           <div className="lg:col-span-7 flex flex-col gap-5">
             <DashboardHero />
-            <InvestimentCard />            
+            <InvestimentCard />
           </div>
-          <BankStatement />         
+          <BankStatement />
         </div>
       )}
     </>
