+++ app/pages/transfers.tsx
@@ -1,9 +1,9 @@
 import { NextSeo } from "next-seo"
-import AsideNav from "@/components/aside-nav"
-import DashboardHero from "@/components/dashboard-hero"
-import BankStatement from "@/components/bank-statement"
+import AsideNav from "@/layouts/blocks/aside-nav"
+import DashboardHero from "@/layouts/blocks/dashboard-hero"
+import BankStatement from "@/layouts/blocks/bank-statement"
 import useStateController from "@/hooks/use-state-controller"
-import TransactionCard from "@/components/transactions"
+import TransactionCard from "@/layouts/blocks/transactions"
 
 export default function TransfersPage() {
   const {authStatus} = useStateController()
