+++ app/utils/date.ts
@@ -1,3 +1,6 @@
+import format from 'date-fns/format'
+import { ptBR } from 'date-fns/locale'
+
 const monthList = [
   "Janeiro",
   "Fevereiro",
@@ -20,18 +23,20 @@ export const formatMonth = (date: Date | string) => {
   return month
 }
 
-export const formatDate = (date: Date | string) => {
+export const formatDate = (date: string | number | Date) => {
   const d = new Date(date)
-  const formattedDate = new Intl.DateTimeFormat('pt-BR', {
-    day: '2-digit',
-    month: '2-digit',
-    year: 'numeric',
-  }).format(d)
-
-  return formattedDate
+  return format(d, 'dd MMM yyyy', {
+    locale:  ptBR,
+  })
 }
 
 const today = new Date()
+
+export const todayFormattedShort =  new Intl.DateTimeFormat('pt-BR', {
+  day: '2-digit',
+  month: '2-digit',
+  year: 'numeric',
+}).format(today)
 
 export const todayFormatted = new Intl.DateTimeFormat('pt-BR', {
   weekday: 'long',
